+++ cli.js
@@ -7,22 +7,28 @@ const indent = (n) => ' '.repeat(n * 4);
 const usage = (cmd, desc) => console.log('\n' + indent(2) + cmd + '\n' + indent(3) + desc.replace(/\n/g, '\n' + indent(4)));
 
 const argOpts = {
-	alias: { full: ['f'], help: ['h'], analyse: ['a', 'analyze'] },
-	boolean: ['full', 'help'],
+	alias: {
+		full: ['f'],
+		help: ['h'],
+		analyse: ['a', 'analyze'],
+		vendored: ['v', 'vendor'],
+	},
+	boolean: ['full', 'help', 'analyze', 'vendored'],
 }
 const args = yargs(process.argv.slice(2), argOpts);
 
 if (args.help) {
 	console.log(`\n${indent(1)}linguist usage:`);
-	usage(`linguist --analyse [<folder>] [--full]`, [
+	usage(`linguist --analyse [<folder>] [--full] [--vendored]`, [
 		`Analyse the language of all files found in a folder.`,
 		`<folder>   The folder to analyse (optional; default './')`,
 		`--full     List every file parsed (optional)`,
+		`--vendored Prevent skipping over vendored files (optional)`,
 	].join('\n'));
 }
 else if (args.analyse) {
 	(async () => {
-		const data = await linguist(args._[0]);
+		const data = await linguist(args._[0], { keepVendored: args.vendored });
 		let loggedData = data;
 		if (!args.full) delete loggedData.results;
 		console.log(loggedData);