+++ changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## Next
+## 1.1.0
 *2021-06-05*
+- Added options argument to analyser function.
+- Added `keepVendored` option to control whether vendored files are kept or not.
 - Added CLI option `--full` to log a full list of parsed files.
+- Added CLI option `--vendored` to include vendored files in output.
+- Fixed input folder not being parsed.
 
 ## 1.0.1
 *2021-06-05*